pyproject.toml
[tool.poetry]
name = "automations"
version = "0.1.0"
description = "One place for all scripts, shared utilities, single config"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0"
python-dotenv = "^1.0"
schedule = "^1.2"
requests = "^2.31"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
black = "^23.0"
ruff = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
